requirements.txt
Flask==1.0.2
python-decouple==3.1
flask-marshmallow==0.9.0
Flask-SQLAlchemy==2.3.2
marshmallow-sqlalchemy==0.16.0
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
psycopg2==2.7.7
psycopg2-binary==2.7.7
zappa==0.47.1
passlib==1.7.1
Flask-JWT-Extended==3.18.0

ipdb==0.11